move QuestionResult into SurveyResults class
